backend/requirements.txt
Django==5.2.9
djangorestframework==3.16.1
python-dotenv==1.2.1
psycopg[binary]==3.3.2
drf-spectacular==0.29.0
django-cors-headers==4.9.0
dj-rest-auth==7.0.1
django-allauth==64.0.0
requests==2.32.5
gunicorn
PyJWT
cryptography
pgvector==0.4.2
elasticsearch==8.11.1
numpy==1.26.4

# 비교함
kiwipiepy==0.22.2
scikit-learn==1.8.0
pandas==2.2.0 # numpy 버전과 호환 고려.
joblib==1.5.3
